Fix connection management in gerrit module
Previously we use a single connection gerrit.API. This works in general
because once HTTPSConnection will reconnect automatically when the
connection is closed, but broken when a request fail and we retry it.

If we sent a request, and getting the response failed, we must close the
connection. If we don't close it, the next request (during retry) will
fail with httplib.CannotSendRequest error, meaning that the connection
state is incorrect.

The easiest way to avoid this issue is to create a new connection for
every request. Since we don't send lot of requests, this is good enough.