Handle transient network connectivity issues
If API call fails with socket.error and NetworkManager reports that we
have connectivity issue, wait until network connectivity is restored or
timeout expires, and retry the function.

The motivation for this is my flaky home ADSL connection. This is
probably my network infrastructure provider fault but I could not get it
fixed yet.

This will be handy to do reliable builds from your favorite cafe or
handle unwanted disconnections when you put your laptop to sleep or
disconnect it temporarily in the way to the next meeting.  In these case
oci will enter network recovery mode and will recover from transient
failures transparently.

Resolves #15.